Split Woo artifact postprocess IO
